refactor(db): 拆分 db 包入口，按职责划分模块
app/db/__init__.py 此前是一个近 800 行的巨型入口，引擎、会话、ORM 基类、事务装饰器、错误诊断挤在一起。本组把它按职责拆成 engine / session / base / decorators / diagnostics 五个模块，__init__ 退化为公开 API 的再导出层；同时让 init.py 名副其实（只做建库/迁移引导），并收口 SQLite URL 推导。

────────────────────────────────────────────────────────────────────────
以下为合并前各原始提交的完整说明（逐字保留其推理与踩坑记录）：
────────────────────────────────────────────────────────────────────────

原始提交 f79d8e39b — refactor(db): 拆分包入口，按职责划分模块

app/db/__init__.py 此前 782 行，是整个 db 模块最大的文件，同时承担六种彼此无关
的职责：驱动错误诊断、引擎构建、会话与连接池、事务装饰器、ORM 基类、连接额度
核算。包入口承载实现导致依赖图难以理清——修连接池时不得不在 monitor 包里写
函数内延迟导入，正是这个结构的连带后果。

按职责拆为 diagnostics / engine / session / decorators / base 五个模块，
__init__.py 收敛为纯 re-export（56 行）。外部 65 个引用文件无需改动。

拆分过程中暴露并修正的问题：
- decorators 实际使用 ScopedSession 而非 SessionFactory，原文件中因同处一个
  命名空间而从未显现
- base 依赖 decorators（Base 的方法带事务装饰器），依赖方向与直觉相反
- connection_budget 依赖 _async_pool_enabled，后者只读配置、归入 engine 层，
  避免 engine 与 session 循环依赖

测试同步适配：原先 patch("app.db.ScopedSession") 打在包入口的 re-export 上，
而装饰器在自己的命名空间按名查找，patch 对其无效——假数据库未生效、真库被写入，
表现为断言数据条数不符而非报错。这类静默失效比 ImportError 危险，已改为
patch 实现模块。

原始提交 41031b620 — refactor(db): init 模块名副其实，收口 SQLite URL 推导

init.py 名为「初始化数据库」，却以 `from app.db import Engine` 在模块级取值——
这意味着它反过来依赖「数据库已在别处初始化完成」。真正的初始化其实发生在
app.db 的 import 副作用里，不由任何叫 init 的代码控制。改为函数内获取引擎，
让创建时机可由调用方决定，这也是后续把连接建立从 import 期移出的前置条件。

SQLite 连接串此前有三处各自拼接（同步引擎、异步引擎、update_db），且 update_db
那处与引擎侧格式不同。两处逻辑一旦漂移，alembic 会连到与应用不同的库文件上，
而这不会报任何错。新增 settings.DB_SQLITE_URL()，与 DB_POSTGRESQL_URL 对称，
三处统一走它。

原始提交 6d8018d3c — refactor(db): 清理拆分残留的未使用导入

按行号切分时带入了几个用不到的符号：base 的 Optional/Tuple、decorators 的
logger（错误日志由 diagnostics 层负责），以及 __init__ 中对私有符号
_register_database_error_logging 的 re-export（无外部引用，不应出现在包接口上）。
